Remove sql_where from plural keys
Closes #25. Multiple statements of sql_where are allowed by Looker's 
validator. 
However, Looker only uses the last statement in the actual SQL join. 
Normally I prefer to follow the behavior of the Looker validator, but in 
this case, I disagree with the way the validator works (it should raise 
an error for 2 statements supplied if only 1 will be used) and will 
allow lkml to raise an error when there are multiple sql_where 
statements in a single join.